[metrics-operator] Add Generic TaskRun histogram
